YUK-15405 Examine Dispatch point loads on registration Delayed point loading until required by postMOAUploadToConnection()

+++ yukon-server/Message/include/msg_ptreg.h
@@ -38,7 +38,7 @@ public:
    virtual CtiMessage* replicateMessage() const;
 
    // If list is empty, I assume you wanted them all!.
-   int  getCount() const;
+   size_t getCount() const;
 
    LONG& operator[](size_t i);
    LONG operator[](size_t i) const;
